di: use standard c++ headers by default
Previously, the default was to define polyfilled versions of c++ headers
and functions, like std::move, std::forward, std::addressof, etc. This
is needed to build the kernel without a c++ standard library, but it is
a poor default, since it forces users to not the c++ standard library
themselves.

Now, the default build uses the standard c++ headers, and this requires
no special configuration. Additionally, this requires passing the
-fsized-deallocation flag to clang, since we use sized deallocation and
need the standard headers to define these prototypes.