ttx: improve terminal emulation
This adds support for color, more renditions, and broken but usable
unicode support.

The new renditions are missing the fancy underline modes (specifically
undercurl), which require better parsing of parameter lists which
mix ; and :.

Unicode support is broken as the current logic assumes a 1:1
relationship between code points and cells. The correct logic is
probably to store the text separately, and calculate the text width
using the unicode data base on each code point. This is also "wrong"
but at least compatible with other terminals.