meta: override dius version for clang Linux CI
The dius version packaged via nix using GCC, and so won't work
when using Clang. The library actually isn't use any std types
for API boundaries, but things still break down because of linking
the wrong standard library.

This might be fixable by not using libc++ over libstdc++ with Clang,
but that would also lessen the value of testing both GCC and Clang.