client: implement renaming a tab via fzf popup
To get the desired behavior with fzf, its necessary to disable some of
the UI components and pass nothing into the input pipe, and additionally
use the --print-query option. The next commit will additionally force
the prompt window to have a height of 3 and be top aligned, which works
perfectly.

Using fzf lets us avoid writing our own line editor just for certain
text prompts. We should be able to use fzf for all the UI components
in ttx.