client+terminal: improve selection experience
The terminal now optimizes damage tracking for the selection which
should improve performance. Additionally, the selection points are
clamped to sane values, which prevents crashes when clicking out
of bounds, and improves behavior when selecting wide cells.

The pane class now renders selection in a prettier manner. It also
handles scrolling the screen while selecting with the mouse properly.
And finally, selection doesn't start until a mouse move event occurs,
which is needed because the selection is inclusive, and otherwise you'd
constantly select characters when trying to use the mouse to change
focus.