client: optimize rendering
The old algorithm basically wrote all "dirty" cells to the screen
unconditionally. This worked terribly when resizing/moving panes,
because that invalidates all cells. It additionally did not account
for overlapping renders (which happens with popups).

The new algorithm maintains 2 terminal::Screen instances, one
representing the current state of the outer terminal, and one
representing the desired state of the outer terminal. The Renderer
now applies changes to the desired screen state, and reconciles the
changes at the end of the render.

This means we only write out cells which have actually changed, even
if the underlying terminal cell thought it changed. This can happen
when a cell is cleared and then redrawn. This also allows us to make
use of the fact that clearing the screen clears the screen (we don't
write out space characters unless necessary).

Changes are grouped in an order designed to reduce the update size,
so all hyperlinks are drawn together, followed by all cells with
identical graphics rendition. Within these groups, cells rendered
from top-to-bottom and left-to-right. We also now try to use a smaller
escape sequence for moving the cursor.

Hopefully these changes make ttx less laggy when used over ssh. And
in general.