terminal+client: begin supporting terminal feature detection
This commit introduces a feature detection mechanism which runs on
client startup. The feature detection involves querying the terminal
and capturing its responses.

Additionally, the relevant query responses are refactored into a
structure with unit tested serialization/deserialization methods.
This is shared between both the actual terminal emulator and the
terminal client.

We'll need to expand support for more feature detection going
forwards.