client: automatically bootstrap terminfo if necessary
On startup we check if our terminfo is available, and if not,
we compile our terminfo to $XDG_STATE_DIR/ttx/terminfo, and
set TERMINFO to point to that directory. This simplifies distribution
and development. Also, for clarity, official distributions should
include the terminfo (and the nix one does), but it still could
be useful to download a static binary we host on GitHub in the future.

Now that we are setting our own TERM variable, there is also an option
to override our default.